# api/openapi.yaml
openapi: 3.1.0
info:
  title: Textile Marketplace MVP API
  version: 0.1.0
  description: Android-first private MVP backend API draft for the textile marketplace.
servers:
  - url: https://api.example.com/v1
    description: Production placeholder
  - url: http://127.0.0.1:8000/v1
    description: Local FastAPI development server
security:
  - bearerAuth: []
...
paths:
...
  /me/saved-items:
    get:
      tags: [Saved]
      summary: List saved items
      operationId: listSavedItems
      parameters:
        - name: target_type
          in: query
          schema:
            $ref: '#/components/schemas/SavedTargetType'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Saved items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedItemsResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Saved]
      summary: Save a profile, work card, or work-needed post
      operationId: saveItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveItemRequest'
      responses:
        '200':
          description: Saved item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedItem'
        default:
          $ref: '#/components/responses/Error'
  /me/saved-items/{saved_item_id}:
    delete:
      tags: [Saved]
      summary: Remove saved item
      operationId: removeSavedItem
      parameters:
        - name: saved_item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Removed
        default:
          $ref: '#/components/responses/Error'
  /reports:
    post:
      tags: [Reports]
      summary: Submit report
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        default:
          $ref: '#/components/responses/Error'
  /me/notifications:
    get:
      tags: [Notifications]
      summary: List in-app notifications
      operationId: listNotifications
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
        default:
          $ref: '#/components/responses/Error'
  /me/notifications/{notification_id}/read:
    post:
      tags: [Notifications]
      summary: Mark notification read
      operationId: markNotificationRead
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        default:
          $ref: '#/components/responses/Error'
  /me/device-tokens:
    post:
      tags: [Notifications]
      summary: Register Android FCM token
      operationId: registerDeviceToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceTokenRequest'
      responses:
        '204':
          description: Registered
        default:
          $ref: '#/components/responses/Error'
  /me/settings:
    get:
      tags: [Settings]
      summary: Get own settings
      operationId: getSettings
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Settings]
      summary: Update own settings
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        default:
          $ref: '#/components/responses/Error'
  /contact-actions:
    post:
      tags: [Contact]
      summary: Log contact action
      operationId: createContactAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactActionRequest'
      responses:
        '202':
          description: Accepted
        default:
          $ref: '#/components/responses/Error'
  /share-links:
    post:
      tags: [Share]
      summary: Create canonical share link and log share event
      operationId: createShareLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareLinkRequest'
      responses:
        '200':
          description: Share link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkResponse'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SavedTargetType:
      type: string
      enum: [profile, work_card, work_needed_post]
    ReportReason:
      type: string
      enum: [wrong_contact, wrong_category, inappropriate_photo, wrong_details, fake_profile, spam, other]
    ReportStatus:
      type: string
      enum: [submitted, in_review, resolved_no_action, action_taken, dismissed]
...
    MediaVisibility:
      type: string
      enum: [public, private_admin_only]
    MediaKind:
      type: string
      enum: [image, document]
    UploadStatus:
      type: string
      enum: [pending_upload, uploaded, processing, ready, failed, deleted]
...
    SearchResult:
      type: object
      required: [result_type, id, profile_id, title, is_verified, photo_count]
      properties:
        result_type:
          type: string
          enum: [profile, work_card, work_needed_post]
        id:
          type: string
          format: uuid
        profile_id:
          type: string
          format: uuid
        title:
          type: string
        subtitle:
          type: string
        category:
          type: string
        product_types:
          type: array
          items:
            type: string
        description:
          type: string
          nullable: true
        locality:
          type: string
          nullable: true
        experience_years:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        is_verified:
          type: boolean
        photo_count:
          type: integer
          minimum: 0
        photos:
          type: array
          items:
            $ref: '#/components/schemas/MediaAsset'
        last_activity_at:
          type: string
          format: date-time
          nullable: true
...
    MediaAsset:
      type: object
      required: [id, media_kind, visibility, upload_status, sort_order]
      properties:
        id:
          type: string
          format: uuid
        media_kind:
          $ref: '#/components/schemas/MediaKind'
        visibility:
          $ref: '#/components/schemas/MediaVisibility'
        upload_status:
          $ref: '#/components/schemas/UploadStatus'
        url:
          type: string
          format: uri
          nullable: true
          description: Public-ready media URL only. Private proof media does not expose URL to owners/public.
        thumbnail_url:
          type: string
          format: uri
        sort_order:
          type: integer
        document_type:
          type: string
        safe_display_name:
          type: string
...
    SavedItemsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SavedItem'
        next_cursor:
          type: string
          nullable: true
    SaveItemRequest:
      type: object
      required: [target_type, target_id]
      properties:
        target_type:
          $ref: '#/components/schemas/SavedTargetType'
        target_id:
          type: string
          format: uuid
    SavedItem:
      type: object
      required: [id, target_type, target_id, profile_role]
      properties:
        id:
          type: string
          format: uuid
        target_type:
          $ref: '#/components/schemas/SavedTargetType'
        target_id:
          type: string
          format: uuid
        profile_role:
          $ref: '#/components/schemas/ProfileRole'
        card:
          $ref: '#/components/schemas/SearchResult'
    CreateReportRequest:
      type: object
      required: [reported_entity_type, reported_entity_id, reason]
      properties:
        reported_entity_type:
          type: string
          enum: [profile, work_card, work_needed_post]
        reported_entity_id:
          type: string
          format: uuid
        reason:
          $ref: '#/components/schemas/ReportReason'
    Report:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ReportStatus'
    NotificationsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        next_cursor:
          type: string
          nullable: true
    Notification:
      type: object
      required: [id, title, message, created_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        message:
          type: string
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
          nullable: true
    RegisterDeviceTokenRequest:
      type: object
      required: [fcm_token, platform]
      properties:
        fcm_token:
          type: string
        platform:
          type: string
          enum: [android]
        device_id:
          type: string
        app_version:
          type: string
    UpdateSettingsRequest:
      type: object
      minProperties: 1
      properties:
        push_notifications_enabled:
          type: boolean
        hidden_from_search:
          type: boolean
    UserSettings:
      type: object
      properties:
        push_notifications_enabled:
          type: boolean
        hidden_from_search:
          type: boolean
    ContactActionRequest:
      type: object
      required: [profile_id, action_type]
      properties:
        profile_id:
          type: string
          format: uuid
        action_type:
          type: string
          enum: [call, whatsapp, address]
        source_type:
          type: string
          enum: [search, saved, profile, work_card, work_needed_post]
        source_id:
          type: string
          format: uuid
    CreateShareLinkRequest:
      type: object
      required: [target_type, target_id]
      properties:
        target_type:
          type: string
          enum: [profile, work_card, work_needed_post]
        target_id:
          type: string
          format: uuid
        channel:
          type: string
          enum: [copy_link, whatsapp, sms, x, email, linkedin, native_other]
    ShareLinkResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        share_text:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            field_errors:
              type: object
              additionalProperties:
                type: string
            request_id:
              type: string